Moved Protocall and Modifed Propagation Event Args
In order to align with the new socket communications, the Propagation
Event Args was modified to use the Protocall.Method enumeration instead
of a string representation of HTTP verbs
