+++ Contracts/Contracts.win.csproj
@@ -83,6 +83,7 @@
     <Compile Include="IRobotProvider.cs" />
     <Compile Include="ISubsystemProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Protocall.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Contracts.win.project.json" />
