+++ Contracts/Events/EventPropagationEventArgs.cs
@@ -7,13 +7,13 @@ namespace Team1922.MVVM.Contracts.Events
 {
     public class EventPropagationEventArgs
     {
-        public EventPropagationEventArgs(string httpMethod, string propertyName, string propertyValue)
+        public EventPropagationEventArgs(Protocall.Method method, string propertyName, string propertyValue)
         {
-            HTTPMethod = httpMethod;
+            Method = method;
             PropertyName = propertyName;
             PropertyValue = propertyValue;
         }
-        public string HTTPMethod { get; }
+        public Protocall.Method Method { get; }
         public string PropertyName { get; set; }
         public string PropertyValue { get; }
     }
