+++ Contracts/Protocall.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Team1922.WebFramework.Sockets
+namespace Team1922.MVVM.Contracts
 {
     public static class Protocall
     {
