+++ WebFramework/Sockets/Request.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Team1922.MVVM.Contracts;
 
 namespace Team1922.WebFramework.Sockets
 {
