+++ WebFramework/WebFramework.win.csproj
@@ -51,7 +51,6 @@
     <Compile Include="Sockets\IRequestDelegator.cs" />
     <Compile Include="Sockets\ISocketClient.cs" />
     <Compile Include="Sockets\ISocketServer.cs" />
-    <Compile Include="Sockets\Protocall.cs" />
     <Compile Include="Sockets\Request.cs" />
     <Compile Include="Sockets\RequestDelegator.cs" />
     <Compile Include="Sockets\RequestResponsePair.cs" />
